Added endpoint and business logic to plan Stories

+++ src/main/java/com/company/product/model/Story.java
@@ -12,7 +12,7 @@ import javax.persistence.Entity;
 public class Story extends Issue {
 
     @Column
-    private int estimatedPointValue;
+    private long estimatedPointValue;
 
     @Column
     private StoryStatus status = StoryStatus.NEW;
