+++ src/main/java/com/company/product/payload/request/StoryRequest.java
@@ -13,7 +13,7 @@ import javax.validation.constraints.Min;
 public class StoryRequest extends IssueRequest {
 
     @Min(value = 1, message = "Estimated point value should not be less than 1")
-    private int estimatedPointValue;
+    private long estimatedPointValue;
 
     private StoryStatus status;
